Move random generators utils to random.h
