+++ src/utils/random.h
@@ -1,0 +1,95 @@
+/*
+ *  Minke is a UCI chess engine
+ *  Copyright (C) 2026 Eduardo Marinho <eduardomarinho@pm.me>
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+#pragma once
+
+#include <cassert>
+#include <cstdint>
+#include <random>
+
+// Implements the Splitmix64 algorithm to generate seeds for the xorshift64star PRNG
+class SeedGenerator {
+  public:
+    static uint64_t master_seed() {
+        // This is not portable considering that random_device may be pseudo-random depending on hardware
+        // but this shouldn't be a problem
+        std::random_device rd{};
+        return static_cast<uint64_t>(rd()) << 32 | static_cast<uint64_t>(rd());
+    }
+
+    SeedGenerator() = delete;
+    SeedGenerator(uint64_t seed = master_seed()) : m_state(seed) {}
+
+    uint64_t next() {
+        m_state += 0x9e3779b97f4a7c15ULL;
+
+        uint64_t z = m_state;
+        z = (z ^ (z >> 30)) * 0xbf58476d1ce4e5b9ULL;
+        z = (z ^ (z >> 27)) * 0x94d049bb133111ebULL;
+        return z ^ (z >> 31);
+    }
+
+  private:
+    uint64_t m_state;
+};
+
+//==== Taken from stockfish
+// xorshift64star Pseudo-Random Number Generator
+// This class is based on original code written and dedicated
+// to the public domain by Sebastiano Vigna (2014).
+// It has the following characteristics:
+//
+//  -  Outputs 64-bit numbers
+//  -  Passes Dieharder and SmallCrush test batteries
+//  -  Does not require warm-up, no zeroland to escape
+//  -  Internal state is a single 64-bit integer
+//  -  Period is 2^64 - 1
+//  -  Speed: 1.60 ns/call (Core i7 @3.40GHz)
+//
+// For further analysis see
+//   <http://vigna.di.unimi.it/ftp/papers/xorshift.pdf>
+class PRNG {
+  public:
+    using result_type = uint64_t;
+
+    PRNG(result_type seed) : s(seed) { assert(seed); }
+
+    template <typename T>
+    T rand() {
+        return T(rand64());
+    }
+
+    // Special generator used to fast init magic numbers.
+    // Output values only have 1/8th of their bits set on average.
+    template <typename T>
+    T sparse_rand() {
+        return T(rand64() & rand64() & rand64());
+    }
+
+    result_type operator()() { return rand64(); }
+    static constexpr result_type min() { return std::numeric_limits<result_type>::min(); }
+    static constexpr result_type max() { return std::numeric_limits<result_type>::max(); }
+
+  public:
+    result_type s;
+
+    result_type rand64() {
+        s ^= s >> 12, s ^= s << 25, s ^= s >> 27;
+        return s * 2685821657736338717LL;
+    }
+};
